changed backLink on license-type step; renamed behaviour;
