+++ apps/common/behaviours/sign-out-on-exit.js
