+++ apps/precursor-chemicals/index.js
@@ -14,7 +14,7 @@ const SetSummaryReferrer = require('../common/behaviours/set-summary-referrer');
 const InformationYouHaveGivenUs = require('../common/behaviours/information-you-have-given-us');
 const SaveFormSession = require('../common/behaviours/save-form-session');
 const ResumeFormSession = require('../common/behaviours/resume-form-session');
-const LogoutOnConfigure = require('../common/behaviours/logout-on-configure');
+const SignOutOnExit = require('../common/behaviours/sign-out-on-exit');
 const Auth = require('../common/behaviours/auth/auth-check');
 
 const steps = {
@@ -520,7 +520,7 @@ const steps = {
   },
 
   '/save-and-exit': {
-    behaviours: [LogoutOnConfigure],
+    behaviours: [SignOutOnExit],
     backLink: false
   }
 };
